move around some code for readability
