+++ scripts/send-packet.js
@@ -12,8 +12,7 @@ const config = require(configPath);
 const sendConfig = config.sendPacket;
 
 const { listenForIbcPacketEvents } = require('./_events.js');
-const { getDispatcher } = require('./_get-vibc-sc.js');
-const { getIbcApp } = require('./_helpers.js');
+const { getDispatcher, getIbcApp } = require('./_vibc-helpers.js');
 
 async function main() {
     const accounts = await hre.ethers.getSigners();
